Move tests folder to be in root rather than in src folder
