+++ package.json
@@ -16,12 +16,12 @@
 		"watch-tslint": "nodemon --exec \"npm run tslint\" -e ts",
 		"debug": "npm run tslint && npm run serve-ts",
 		"watch-debug": "./node_modules/concurrently/bin/concurrently.js -k -p \"[{name}]\" -n \"Lint,Run\" -c \"yellow.bold,cyan.bold\" \"npm run watch-tslint\" \"npm run watch-ts\"",
-		"test+cov": "./node_modules/nyc/bin/nyc.js --reporter=text ./node_modules/mocha/bin/mocha --require ts-node/register --require source-map-support/register --full-trace --exit src/tests/**/*.ts src/tests/*.ts",
-		"test": "./node_modules/mocha/bin/mocha --require ts-node/register --require source-map-support/register --full-trace --exit src/tests/**/*.ts src/tests/*.ts",
+		"test+cov": "./node_modules/nyc/bin/nyc.js --reporter=text ./node_modules/mocha/bin/mocha --require ts-node/register --require source-map-support/register --full-trace --exit tests/**/*.ts tests/*.ts",
+		"test": "./node_modules/mocha/bin/mocha --require ts-node/register --require source-map-support/register --full-trace --exit tests/**/*.ts tests/*.ts",
 		"test+cov-some": "./node_modules/nyc/bin/nyc.js --reporter=text ./node_modules/mocha/bin/mocha --require ts-node/register --require source-map-support/register --full-trace --exit",
 		"test-some": "./node_modules/mocha/bin/mocha --require ts-node/register --require source-map-support/register --full-trace --exit",
-		"test-landing": "./node_modules/mocha/bin/mocha --require ts-node/register --require source-map-support/register --full-trace --exit src/tests/**/*.ts src/tests/*.ts --reporter landing",
-		"test-nyan": "./node_modules/mocha/bin/mocha --require ts-node/register --require source-map-support/register --full-trace --exit src/tests/**/*.ts src/tests/*.ts --reporter nyan",
+		"test-landing": "./node_modules/mocha/bin/mocha --require ts-node/register --require source-map-support/register --full-trace --exit tests/**/*.ts tests/*.ts --reporter landing",
+		"test-nyan": "./node_modules/mocha/bin/mocha --require ts-node/register --require source-map-support/register --full-trace --exit tests/**/*.ts tests/*.ts --reporter nyan",
 		"watch-test+cov": "nodemon -e ts --exec \"npm run test+cov\"",
 		"watch-test": "nodemon -e ts --exec \"npm run test\"",
 		"watch-test+cov-some": "nodemon -e ts --exec \"npm run test+cov-some\"",
