+++ tsconfig.json
@@ -12,7 +12,9 @@
     },
     "include": [
 		"src/*",
-        "src/**/*"
+        "src/**/*",
+		"tests/*",
+		"tests/**/*"
     ],
 	"exclude": [
 		"node_modules"
